test: tsd 테스트를 jest 테스트로 수정한다.
