+++ __tests__/type.test.ts
@@ -1,0 +1,56 @@
+import _ from "../src";
+import { expectType } from "tsd";
+import { FetchOptions, FetchResponse } from "../src/util";
+
+test("fetch 함수에 대한 타입을 체크한다.", () => {
+  // given
+  type TestData = {
+    id: number;
+    name: string;
+    age: number;
+  };
+
+  expectType<
+    (url: string, options?: FetchOptions) => Promise<FetchResponse<TestData[]>>
+  >(_.fetch<TestData[]>);
+});
+
+test("isNull 함수에 대한 타입을 체크한다.", () => {
+  expectType<(value: any) => value is null>(_.isNull);
+});
+
+test("isNill 함수에 대한 타입을 체크한다.", () => {
+  expectType<(value: any) => value is null | undefined>(_.isNil);
+});
+
+test("isFunction 함수에 대한 타입을 체크한다.", () => {
+  expectType<(value: any) => value is (...args: any[]) => any>(_.isFunction);
+});
+
+test("shuffle 함수에 대한 타입을 체크한다.", () => {
+  expectType<(array: number[]) => number[]>(_.shuffle<number>);
+});
+
+test("pick 함수에 대한 타입을 체크한다.", () => {
+  // given
+  type testObj = { a: 1; b: 2; c: 3 };
+  type testPath = ["a", "b"];
+  type pickResultObj = { a: 1; b: 2 };
+
+  // then
+  expectType<(object: testObj, path: testPath) => pickResultObj>(
+    _.pick<testObj, testPath>
+  );
+});
+
+test("omit 함수에 대한 타입을 체크한다.", () => {
+  // given
+  type testObj = { a: 1; b: 2; c: 3 };
+  type testPath = ["a", "b"];
+  type OmitResultObj = { c: 3 };
+
+  // then
+  expectType<(object: testObj, path: testPath) => OmitResultObj>(
+    _.omit<testObj, testPath>
+  );
+});
